requirements.txt.txt
python-dotenv==1.0.1
aiogram==3.1.4
